Extract src/components/Attendees/usePasteHandler.ts
