+++ src/components/Attendees/PeopleSearch.tsx
@@ -27,6 +27,7 @@ import {
 import { useI18n } from "twake-i18n";
 import { ResourceIcon } from "./ResourceIcon";
 import { isValidEmail } from "../../utils/isValidEmail";
+import { usePasteHandler } from "./usePasteHandler";
 
 export interface User {
   email: string;
@@ -124,55 +125,14 @@ export function PeopleSearch({
     [query, selectedUsers, onChange, t, setInputError, setQuery]
   );
 
-  const handlePaste = useCallback(
-    (event: React.ClipboardEvent<HTMLInputElement>) => {
-      if (!freeSolo) return;
-
-      const pasted = event.clipboardData.getData("text/plain");
-      if (!pasted) return;
-
-      // Split by comma, semicolon, newline, or whitespace
-      const chunks = pasted
-        .split(/[,;\n\r\s]+/)
-        .map((s) => s.trim())
-        .filter(Boolean);
-
-      // If there is only one chunk, let the default Autocomplete behaviour handle it
-      if (chunks.length <= 1) return;
-
-      event.preventDefault();
-
-      const existingEmails = new Set(selectedUsers.map((u) => u.email));
-      const validUsers: User[] = [];
-      const invalid: string[] = [];
-
-      for (const chunk of chunks) {
-        if (!isValidEmail(chunk)) {
-          invalid.push(chunk);
-        } else if (!existingEmails.has(chunk)) {
-          existingEmails.add(chunk);
-          validUsers.push({ email: chunk, displayName: chunk });
-        }
-        // silently skip duplicates
-      }
-
-      if (validUsers.length > 0) {
-        onChange(event, [...selectedUsers, ...validUsers]);
-      }
-
-      if (invalid.length > 0) {
-        // Leave the invalid text in the input for manual correction
-        setQuery(invalid.join(", "));
-        setInputError(
-          t("peopleSearch.invalidEmail").replace("%{email}", invalid.join(", "))
-        );
-      } else {
-        setQuery("");
-        setInputError(null);
-      }
-    },
-    [freeSolo, selectedUsers, onChange, setQuery, setInputError, t]
-  );
+  const handlePaste = usePasteHandler({
+    freeSolo,
+    selectedUsers,
+    onChange,
+    setQuery,
+    setInputError,
+    t,
+  });
 
   const defaultRenderInput = useCallback(
     (params: AutocompleteRenderInputParams) => {
